requirements.txt
pyTelegramBotAPI
Flask
python-dotenv
gunicorn
